backend/requirements.txt
# Dépendances Python pour le système multiagent pharma
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0

# LLM Providers
google-generativeai==0.3.2
ollama>=0.1.8

# Base de données et stockage
aiosqlite==0.19.0
redis==5.0.1

# Logging et monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Utilitaires
python-multipart==0.0.6
python-dotenv==1.0.0
pyyaml==6.0.1
httpx>=0.25.2
aiohttp==3.9.1

# Analyse de texte et NLP
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3

# Validation et sécurité
cryptography==41.0.7
bcrypt==4.1.1

# Tests
pytest==7.4.3
pytest-asyncio==0.21.1

# Développement
black==23.11.0
isort==5.12.0
flake8==6.1.0